+++ src/Common/crypto/mbedtls/key_manager_lib.c
@@ -265,7 +265,8 @@ SOPC_ReturnStatus SOPC_KeyManager_Certificate_CreateOrAddFromDER(const uint8_t* 
         if (0 != err)
         {
             status = SOPC_STATUS_NOK;
-            fprintf(stderr, "> KeyManager: certificate buffer parse failed with error code: -0x%X\n", -err);
+            fprintf(stderr, "> KeyManager: certificate buffer parse failed with error code: -0x%X\n",
+                    (unsigned int) -err);
         }
     }
 
@@ -308,7 +309,7 @@ SOPC_ReturnStatus SOPC_KeyManager_Certificate_CreateOrAddFromFile(const char* sz
         {
             status = SOPC_STATUS_NOK;
             fprintf(stderr, "> KeyManager: certificate file \"%s\" parse failed with error code: -0x%X\n", szPath,
-                    -err);
+                    (unsigned int) -err);
         }
     }
 
@@ -712,7 +713,7 @@ static char* get_raw_sha1(const mbedtls_x509_buf* raw)
     int err = mbedtls_md(pmd, raw->p, raw->len, pDest);
     if (0 != err)
     {
-        fprintf(stderr, "Cannot compute thumbprint of certificate, err -0x%X\n", -err);
+        fprintf(stderr, "Cannot compute thumbprint of certificate, err -0x%X\n", (unsigned int) -err);
         return NULL;
     }
 
@@ -925,7 +926,7 @@ SOPC_ReturnStatus SOPC_KeyManager_CRL_CreateOrAddFromDER(const uint8_t* bufferDE
         if (0 != err)
         {
             status = SOPC_STATUS_NOK;
-            fprintf(stderr, "> KeyManager: crl buffer parse failed with error code: -0x%X\n", -err);
+            fprintf(stderr, "> KeyManager: crl buffer parse failed with error code: -0x%X\n", (unsigned int) -err);
         }
     }
 
@@ -957,7 +958,8 @@ SOPC_ReturnStatus SOPC_KeyManager_CRL_CreateOrAddFromFile(const char* szPath, SO
         if (0 != err)
         {
             status = SOPC_STATUS_NOK;
-            fprintf(stderr, "> KeyManager: crl file \"%s\" parse failed with error code: -0x%X", szPath, -err);
+            fprintf(stderr, "> KeyManager: crl file \"%s\" parse failed with error code: -0x%X", szPath,
+                    (unsigned int) -err);
         }
     }
